requirements.txt
fastapi
uvicorn
twilio
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok